Revert "test: update struct"
This reverts commit b97c2b76dcece031820bad1fc2f8ac3b72fdeb1c.

+++ examples/http/main.c
@@ -10,7 +10,7 @@ struct http_parser_state_s {
   void* current;
   int error;
   const char* reason;
-  uint8_t index;
+  int index;
   void* data;
 
   uint8_t method;
